added userform settings

+++ project.front/src/app/components/dashboard/dashboard.component.ts
@@ -33,15 +33,18 @@ export class DashboardComponent implements OnInit {
   }
 
 
-  deleteUser(id: number): void {
-    this.usersService.deleteData(id).subscribe({
-      next: () => {
-        this.userList = this.userList.filter(user => user.id !== id);
-      },
-      error: (err) => {
-        console.error('Error al eliminar el usuario', err);
-      }
-    });
+  deleteUser(id?: number): void {
+    if (id !== undefined) {
+      this.usersService.deleteData(id).subscribe({
+        next: () => {
+          this.userList = this.userList.filter(user => user.id !== id);
+        },
+        error: (err) => {
+          console.error('Error al eliminar el usuario', err);
+        }
+      });
+    } else {
+      console.error('No se proporcionó un ID de usuario');
+    }
   }
-  
 }
